Api response modified for  different time frame
