+++ backend/Python Modules/Contest-redis/app.py
@@ -3,37 +3,52 @@ from flask import Flask, request, jsonify
 import requests
 import json
 # Make Redis
+from datetime import datetime
 redis_cache = redis.Redis(host='localhost', port=6379, db=0)
 
 # Make Flask app
 app = Flask(__name__)
+def is_contest_time():
+    # Get the current time in Coordinated Universal Time (UTC)
+    current_time_in_utc = datetime.utcnow()
 
+    # Check if it's Sunday and between 2:30 AM and 4:00 AM UTC
+    if current_time_in_utc.weekday() == 6:  # Sunday is represented as 6 in Python
+        if (current_time_in_utc.hour == 13 and current_time_in_utc.minute >= 30) or \
+           (current_time_in_utc.hour == 14) or \
+           (current_time_in_utc.hour == 15 and current_time_in_utc.minute == 0):
+            return True
+
+    return False
 @app.route('/get-prediction', methods=['GET'])
 def get_prediction():
-    weekly_contest = request.args.get('weekly_contest', 'weekly-contest-367')
-    username = request.args.get('username', 'sahapriyam')
-    
-    
-    cache_key = f'{weekly_contest}-{username}'
-    
-   
-    cached_data = redis_cache.get(cache_key)
-    
-    if cached_data:
-        
-        cached_data = json.loads(cached_data)
-        return jsonify({'source': 'Redis', 'data': cached_data})
+    if is_contest_time():
+        return "Contest is running."
     else:
-        api_url = f'https://lccn.lbao.site/api/v1/contest-records/user?contest_name={weekly_contest}&username={username}&archived=false'
-
-        response = requests.get(api_url)
-
-        if response.status_code == 200:
-            data = response.json()
-            redis_cache.setex(cache_key, 3600, json.dumps(data))
-            return jsonify({'source': 'API', 'data': data})
+        weekly_contest = request.args.get('weekly_contest', 'weekly-contest-367')
+        username = request.args.get('username', 'sahapriyam')
+        
+        
+        cache_key = f'{weekly_contest}-{username}'
+        
+    
+        cached_data = redis_cache.get(cache_key)
+        
+        if cached_data:
+            
+            cached_data = json.loads(cached_data)
+            return jsonify({'source': 'Redis', 'data': cached_data})
         else:
-            return jsonify({'source': 'Error', 'error': 'Failed to fetch data from the API'}), 500
+            api_url = f'https://lccn.lbao.site/api/v1/contest-records/user?contest_name={weekly_contest}&username={username}&archived=false'
+
+            response = requests.get(api_url)
+
+            if response.status_code == 200:
+                data = response.json()
+                redis_cache.setex(cache_key, 3600, json.dumps(data))
+                return jsonify({'source': 'API', 'data': data})
+            else:
+                return jsonify({'source': 'Error', 'error': 'Failed to fetch data from the API'}), 500
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0', port=8080, debug=True)
